requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
google-generativeai==0.7.2
google-adk==0.1.0
sqlalchemy==2.0.30
asyncpg==0.31.0
psycopg2-binary==2.9.11
google-auth>=2.48.1,<3.0.0
google-auth-oauthlib==1.2.0
google-api-python-client >= 2.157.0
pydantic==2.12.5
python-dotenv==1.0.1
greenlet==3.3.2